+++ chalice/__init__.py
@@ -5,4 +5,4 @@ from chalice.app import (
     CustomAuthorizer, CognitoUserPoolAuthorizer
 )
 
-__version__ = '0.8.0'
+__version__ = '0.8.1'